+++ service/controller/resource/legacyfinalizer/delete.go
@@ -1,0 +1,8 @@
+package legacyfinalizer
+
+import "context"
+
+// EnsureDeleted is no-op.
+func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
+	return nil
+}
